Create a generic component to update and insert Items
