--- frontend/app/src/scenes/Home/scenes/AddItem/interfaces.ts
@@ -1,10 +1,0 @@
-import { SubCategory } from '../../interfaces';
-
-export interface AddItemStateProps {
-  isLoading: boolean;
-  subCategories: SubCategory[];
-}
-
-export interface AddItemPayload {
-  id: number;
-}
