--- frontend/app/src/scenes/Home/scenes/EditItem/interfaces.ts
@@ -1,10 +1,0 @@
-import { Item } from '../../interfaces';
-
-export interface EditItemStateProps {
-  isLoading: boolean;
-  item: Item | null;
-}
-
-export interface ItemUrlParams {
-  id: string;
-}
