+++ frontend/app/src/scenes/Home/scenes/UpsertItem/interfaces.ts
@@ -1,0 +1,11 @@
+import { Item, SubCategory } from '../../interfaces';
+
+export interface AddItemStateProps {
+  isLoading: boolean;
+  subCategories: SubCategory[];
+  item: Item | null;
+}
+
+export interface ItemUrlParams {
+  itemId: string;
+}
